refactor(lockfile): split bun modules
